--- a/benchmarks/Makefile
+++ b/benchmarks/Makefile
@@ -1,7 +1,7 @@
 
 bench:
 	# @go test -bench BenchmarkMarshalJSON -benchmem -memprofile mem.out -cpuprofile cpu.out -run NONE
-	@go test -bench BenchmarkParse -benchmem -count=10 -run NONE > new.txt
+	@go test -bench BenchmarkStringCmp -benchmem -count=10 -run NONE > new.txt
 
 bench-udec:
 	@rm -f bench-udec.txt